move Fault into utils
